+++ src/js/background-script.js
@@ -93,6 +93,9 @@ const listenerRuntime = function(request, sender, sendResponse) {
   else if (request.messageId === "Register") {
     registerAPI(request, sender.id, sendResponse);
   }
+  else if (request.messageId === "CheckUrl") {
+    checkUrlAPI(request, sender.id, sendResponse);
+  }
   else if (request.messageId === "GetSession") {
     sendResponse({
       userMail: coinformUserMail,
@@ -350,6 +353,21 @@ const registerAPI = function(request, scriptId, registerCallback) {
   client.postUserRegister(request.userMail, request.userPass).then(res => registerCallback(res)).catch(err => {
     logger.logMessage(CoInformLogger.logTypes.error, `Request error: ${err}`, scriptId);
     // console.error(err);
+  });
+
+};
+
+const checkUrlAPI = function(request, scriptId, checkUrlCallback) {
+
+  logger.logMessage(CoInformLogger.logTypes.debug, `Check URL: ${request.url}`, scriptId);
+
+  client.getCheckUrlInfo(request.url).then(res => checkUrlCallback(res)).catch(err => {
+    logger.logMessage(CoInformLogger.logTypes.error, `Request error: ${err}`, scriptId);
+    // console.error(err);
+    if (checkUrlCallback) checkUrlCallback({
+      status: -1,
+      error: err
+    });
   });
 
 };