+++ src/js/coinform-client.js
@@ -18,6 +18,14 @@ CoinformClient.prototype = {
     return getResponseTweet(this.baseURL + '/response/' + queryID.replace(/['"]+/g, ''));
   },
 
+  getCheckUrlInfo: function (url) {
+    return getCheckUrl(this.baseURL + '/check-url/', url);
+  },
+
+  postTwitterEvaluate: function (tweetId, evaluation, userToken) {
+    return postEvaluate(this.baseURL + '/twitter/evaluate/', tweetId, evaluation, userToken);
+  },
+
   postTwitterEvaluate: function (tweetId, tweetUrl, evaluation, userToken) {
     return postEvaluate(this.baseURL + '/twitter/evaluate/', tweetId, tweetUrl, evaluation, userToken);
   },
@@ -138,6 +146,46 @@ function getResponseTweet(path) {
       .catch(err => reject(err));
   });
 
+}
+
+function getCheckUrl(path, url) {
+
+  const data = {source: url};
+  const urlParams = buildUrl(path, data);
+
+  return new Promise((resolve, reject) => {
+    f(urlParams, {
+      method: 'GET',
+      // mode: 'cors',
+      headers: {
+        'Accept': 'application/json',
+        'Connection': 'keep-alive',
+        'rejectUnauthorized': false
+      }
+    })
+      .then(res => res.json().then(json => ({
+        status: res.status,
+        data: json
+      })))
+      .then(res => resolve(res))
+      .catch(err => reject(err));
+  });
+
+}
+
+function buildUrl(url, parameters) {
+  let qs = "";
+  for (const key in parameters) {
+    if (parameters.hasOwnProperty(key)) {
+      const value = parameters[key];
+      qs += encodeURIComponent(key) + "=" + encodeURIComponent(value) + "&";
+    }
+  }
+  if (qs.length > 0) {
+    qs = qs.substring(0, qs.length - 1); //chop off last "&"
+    url = url + "?" + qs;
+  }
+  return url;
 }
 
 function getHttpRequest(path) {